Simplify HTML
Impressed by yesterday's presentation tried to make the markup simple
and understandable.
Fixed strange errors (-: Improved structure. Verified for compliance
with the perfect criterions. Remove all classes, except for the
sections, and many other attributes which are not necessary yet, in
particular in the forms. Also deleted additional containers.
All of this done in order to better perceive the markup and avoid
confusion, when'll deal with stylization.
Names of classes and additional containers will create when'll begin to
stylization.